Refactor: Move spelling answer endpoint to RecallPromptController
