+++ backend/src/main/java/com/odde/doughnut/controllers/RecallPromptController.java
@@ -2,9 +2,13 @@ package com.odde.doughnut.controllers;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.odde.doughnut.controllers.dto.AnswerDTO;
+import com.odde.doughnut.controllers.dto.AnswerSpellingDTO;
 import com.odde.doughnut.controllers.dto.QuestionContestResult;
+import com.odde.doughnut.controllers.dto.SpellingResultDTO;
 import com.odde.doughnut.entities.*;
+import com.odde.doughnut.exceptions.UnexpectedNoAccessRightException;
 import com.odde.doughnut.services.AuthorizationService;
+import com.odde.doughnut.services.MemoryTrackerService;
 import com.odde.doughnut.services.RecallQuestionService;
 import com.odde.doughnut.testability.TestabilitySettings;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -21,23 +25,18 @@ class RecallPromptController {
 
   private final RecallQuestionService recallQuestionService;
   private final AuthorizationService authorizationService;
+  private final MemoryTrackerService memoryTrackerService;
 
   @Autowired
   public RecallPromptController(
       RecallQuestionService recallQuestionService,
       TestabilitySettings testabilitySettings,
-      AuthorizationService authorizationService) {
+      AuthorizationService authorizationService,
+      MemoryTrackerService memoryTrackerService) {
     this.testabilitySettings = testabilitySettings;
     this.authorizationService = authorizationService;
     this.recallQuestionService = recallQuestionService;
-  }
-
-  @GetMapping("/{memoryTracker}/question")
-  @Transactional
-  public RecallPrompt askAQuestion(
-      @PathVariable("memoryTracker") @Schema(type = "integer") MemoryTracker memoryTracker) {
-    authorizationService.assertLoggedIn();
-    return recallQuestionService.generateAQuestion(memoryTracker);
+    this.memoryTrackerService = memoryTrackerService;
   }
 
   @PostMapping("/{recallPrompt}/regenerate")
@@ -69,6 +68,21 @@ class RecallPromptController {
       @Valid @RequestBody AnswerDTO answerDTO) {
     authorizationService.assertLoggedIn();
     return recallQuestionService.answerQuestion(
+        recallPrompt,
+        answerDTO,
+        authorizationService.getCurrentUser(),
+        testabilitySettings.getCurrentUTCTimestamp());
+  }
+
+  @PostMapping("/{recallPrompt}/answer-spelling")
+  @Transactional
+  public SpellingResultDTO answerSpelling(
+      @PathVariable("recallPrompt") @Schema(type = "integer") RecallPrompt recallPrompt,
+      @Valid @RequestBody AnswerSpellingDTO answerDTO)
+      throws UnexpectedNoAccessRightException {
+    authorizationService.assertLoggedIn();
+    authorizationService.assertReadAuthorization(recallPrompt.getMemoryTracker());
+    return memoryTrackerService.answerSpelling(
         recallPrompt,
         answerDTO,
         authorizationService.getCurrentUser(),
