+++ backend/src/main/java/com/odde/doughnut/controllers/dto/AnswerSpellingDTO.java
@@ -6,5 +6,4 @@ import lombok.Setter;
 public class AnswerSpellingDTO {
   @Getter @Setter private String spellingAnswer;
   @Getter @Setter private Integer thinkingTimeMs;
-  @Getter @Setter private Integer recallPromptId;
 }
