+++ frontend/tests/helpers/index.ts
@@ -69,8 +69,9 @@ const methodToController: Record<string, any> = {
   regenerate: RecallPromptController,
   contest: RecallPromptController,
   answerQuiz: RecallPromptController,
+  answerSpelling: RecallPromptController,
   showQuestion: RecallPromptController,
-  askAQuestion: RecallPromptController,
+  askAQuestion: MemoryTrackerController,
   toggleApproval: PredefinedQuestionController,
   suggestQuestionForFineTuning: PredefinedQuestionController,
   refineQuestion: PredefinedQuestionController,
@@ -115,10 +116,8 @@ const methodToController: Record<string, any> = {
   selfEvaluate: MemoryTrackerController,
   removeFromRepeating: MemoryTrackerController,
   reEnable: MemoryTrackerController,
-  answerSpelling: MemoryTrackerController,
   markAsRepeated: MemoryTrackerController,
   showMemoryTracker: MemoryTrackerController,
-  getSpellingQuestion: MemoryTrackerController,
   getRecentlyReviewed: MemoryTrackerController,
   getRecentMemoryTrackers: MemoryTrackerController,
   getRecallPrompts: MemoryTrackerController,
