+++ open_api_docs.yaml
@@ -668,6 +668,36 @@ paths:
             '*/*':
               schema:
                 $ref: "#/components/schemas/AnsweredQuestion"
+  /api/recall-prompts/{recallPrompt}/answer-spelling:
+    post:
+      tags:
+      - recall-prompt-controller
+      operationId: answerSpelling
+      parameters:
+      - name: recallPrompt
+        in: path
+        required: true
+        schema:
+          type: integer
+      requestBody:
+        content:
+          application/json:
+            schema:
+              $ref: "#/components/schemas/AnswerSpellingDTO"
+        required: true
+      responses:
+        "500":
+          description: Internal Server Error
+          content:
+            '*/*':
+              schema:
+                type: string
+        "200":
+          description: OK
+          content:
+            '*/*':
+              schema:
+                $ref: "#/components/schemas/SpellingResultDTO"
   /api/predefined-questions/{predefinedQuestion}/toggle-approval:
     post:
       tags:
@@ -1425,36 +1455,6 @@ paths:
             '*/*':
               schema:
                 $ref: "#/components/schemas/MemoryTracker"
-  /api/memory-trackers/{memoryTracker}/answer-spelling:
-    post:
-      tags:
-      - memory-tracker-controller
-      operationId: answerSpelling
-      parameters:
-      - name: memoryTracker
-        in: path
-        required: true
-        schema:
-          type: integer
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: "#/components/schemas/AnswerSpellingDTO"
-        required: true
-      responses:
-        "500":
-          description: Internal Server Error
-          content:
-            '*/*':
-              schema:
-                type: string
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/SpellingResultDTO"
   /api/mcp/notes/create:
     post:
       tags:
@@ -2652,30 +2652,6 @@ paths:
             '*/*':
               schema:
                 $ref: "#/components/schemas/DueMemoryTrackers"
-  /api/recall-prompts/{memoryTracker}/question:
-    get:
-      tags:
-      - recall-prompt-controller
-      operationId: askAQuestion
-      parameters:
-      - name: memoryTracker
-        in: path
-        required: true
-        schema:
-          type: integer
-      responses:
-        "500":
-          description: Internal Server Error
-          content:
-            '*/*':
-              schema:
-                type: string
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/RecallPrompt"
   /api/predefined-questions/{note}/export-question-generation:
     get:
       tags:
@@ -2910,30 +2886,6 @@ paths:
             '*/*':
               schema:
                 $ref: "#/components/schemas/MemoryTracker"
-  /api/memory-trackers/{memoryTracker}/spelling-question:
-    get:
-      tags:
-      - memory-tracker-controller
-      operationId: getSpellingQuestion
-      parameters:
-      - name: memoryTracker
-        in: path
-        required: true
-        schema:
-          type: integer
-      responses:
-        "500":
-          description: Internal Server Error
-          content:
-            '*/*':
-              schema:
-                type: string
-        "200":
-          description: OK
-          content:
-            '*/*':
-              schema:
-                $ref: "#/components/schemas/RecallPrompt"
   /api/memory-trackers/{memoryTracker}/recall-prompts:
     get:
       tags:
@@ -2960,6 +2912,30 @@ paths:
                 type: array
                 items:
                   $ref: "#/components/schemas/RecallPrompt"
+  /api/memory-trackers/{memoryTracker}/question:
+    get:
+      tags:
+      - memory-tracker-controller
+      operationId: askAQuestion
+      parameters:
+      - name: memoryTracker
+        in: path
+        required: true
+        schema:
+          type: integer
+      responses:
+        "500":
+          description: Internal Server Error
+          content:
+            '*/*':
+              schema:
+                type: string
+        "200":
+          description: OK
+          content:
+            '*/*':
+              schema:
+                $ref: "#/components/schemas/RecallPrompt"
   /api/memory-trackers/recently-reviewed:
     get:
       tags:
@@ -3849,6 +3825,28 @@ components:
       - answer
       - predefinedQuestion
       - recallPromptId
+    AnswerSpellingDTO:
+      type: object
+      properties:
+        spellingAnswer:
+          type: string
+        thinkingTimeMs:
+          type: integer
+          format: int32
+    SpellingResultDTO:
+      type: object
+      properties:
+        note:
+          $ref: "#/components/schemas/Note"
+        answer:
+          type: string
+        isCorrect:
+          type: boolean
+        memoryTrackerId:
+          type: integer
+          format: int32
+      required:
+      - note
     QuestionSuggestionCreationParams:
       type: object
       properties:
@@ -4015,31 +4013,6 @@ components:
       required:
       - id
       - nextRecallAt
-      - note
-    AnswerSpellingDTO:
-      type: object
-      properties:
-        spellingAnswer:
-          type: string
-        thinkingTimeMs:
-          type: integer
-          format: int32
-        recallPromptId:
-          type: integer
-          format: int32
-    SpellingResultDTO:
-      type: object
-      properties:
-        note:
-          $ref: "#/components/schemas/Note"
-        answer:
-          type: string
-        isCorrect:
-          type: boolean
-        memoryTrackerId:
-          type: integer
-          format: int32
-      required:
       - note
     McpNoteAddDTO:
       type: object
